requirements.txt
Flask==3.1.1
flask-cors==6.0.1
google-generativeai==0.8.5
